% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globV.R
\docType{data}
\name{control.default}
\alias{control.default}
\title{Global variable of spatial variable selection, contains optimization tuning
parameters.}
\format{\describe{
\item{\code{maxIter}}{maximum number of iterations in the \link{SpatialVS}}
\item{\code{iwls}}{convergence criterion for the iterative weighted least squares}
\item{\code{tol1}}{convergence criteria for beta estimation in \link{SpatialVS}}
\item{\code{tol2}}{convergence criteria for theta estimation in \link{SpatialVS}}
}}
\usage{
control.default
}
\description{
control.default is a list that gives the default setting for the optimization procedures
}
\examples{
control.default=list(maxIter=200,iwls=10^(-4),tol1=10^(-3),tol2=10^(-3))
}
\keyword{default setting}
