% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{SpatialVS.summary}
\alias{SpatialVS.summary}
\title{Function for spatial variable selection's summary}
\usage{
SpatialVS.summary(obj)
}
\arguments{
\item{obj}{List, returned by \link{SpatialVS}.}
}
\value{
A vector contains the final estimates of parameters. The estimates of theta are in log scale.
}
\description{
return the summarized parameter estimates from the \link{SpatialVS} procedure.
}
\examples{

test.fit<-SpatialVS(dat.obj=small.test.dat, alpha.vec=0.5, lambda.vec=5,
method="PQL", intercept=TRUE, verbose=FALSE)
SpatialVS.summary(test.fit)
}
